Move to plugins.py
