+++ qiime2/sdk/util.py
@@ -183,16 +183,3 @@ def get_available_usage_drivers() -> Dict[str, 'UsageDriver']:
         entry_point.name: entry_point.resolve() for entry_point in
         iter_entry_points(group='qiime2.usage_drivers')
     }
-
-
-def assert_no_nans_in_tables(fh):
-    '''
-    Checks for NaNs present in any of the tables in the indicated file then
-    resets to the head of the file.
-    '''
-    from pandas import read_html
-
-    tables = read_html(fh)
-    for df in tables:
-        assert not df.isnull().values.any()
-    fh.seek(0)
